Separate the main javascript code into another file
