+++ popup.html
@@ -30,6 +30,6 @@
       </div>
     </div>
 
-    <script src="popup.js"></script>
+    <script src="main.js"></script>
   </body>
 </html>
